+++ lib/image/options/crop.ex
@@ -4,6 +4,13 @@ defmodule Image.Options.Crop do
 
   """
 
+  import Image,
+    only: [
+      is_percent: 1,
+      is_positive_percent: 1,
+      is_box: 4
+    ]
+
   @typedoc """
   Options applicable to cropping an
   image.
@@ -64,6 +71,10 @@ defmodule Image.Options.Crop do
 
   @crop Map.keys(@crop_map)
 
+  @doc """
+  Validates options to `Iamge.crop/5`.
+
+  """
   def validate_options(options) do
     case Enum.reduce_while(options, options, &validate_option(&1, &2)) do
       {:error, value} ->
@@ -74,7 +85,7 @@ defmodule Image.Options.Crop do
     end
   end
 
-  def validate_option(_other, options) do
+  defp validate_option(_other, options) do
     {:cont, options}
   end
 
@@ -86,6 +97,57 @@ defmodule Image.Options.Crop do
 
   def validate_crop(crop, _options) do
     {:halt, {:error, invalid_crop(crop)}}
+  end
+
+  @doc false
+  def normalize_box({w, _h} = dims, left, top, width, height) when is_percent(left) do
+    normalize_box(dims, round(left * w), top, width, height)
+  end
+
+  def normalize_box({_w, h} = dims, left, top, width, height) when is_percent(top) do
+    normalize_box(dims, left, round(top * h), width, height)
+  end
+
+  def normalize_box({w, _h} = dims, left, top, width, height) when is_positive_percent(width) do
+    normalize_box(dims, left, top, round(width * w), height)
+  end
+
+  def normalize_box({_w, h} = dims, left, top, width, height) when is_positive_percent(height) do
+    normalize_box(dims, left, top, width, round(height * h))
+  end
+
+  def normalize_box(_dims, left, top, width, height) when is_box(left, top, width, height) do
+    {left, top, width, height}
+  end
+
+  def normalize_box(_dims, _left, _top, width, _height)
+      when not is_integer(width) and not is_positive_percent(width) do
+    {:error, size_error("width", width)}
+  end
+
+  def normalize_box(_dims, _left, _top, _width, height)
+      when not is_integer(height) and not is_positive_percent(height) do
+    {:error, size_error("height", height)}
+  end
+
+  def normalize_box(_dims, left, _top, _width, _height)
+      when not is_integer(left) and not is_positive_percent(left) do
+    {:error, location_error("left", left)}
+  end
+
+  def normalize_box(_dims, _left, top, _width, _height)
+      when not is_integer(top) and not is_positive_percent(top) do
+    {:error, location_error("top", top)}
+  end
+
+  defp size_error(dim, size) do
+    "#{dim} must be a percentage expressed as a float greater than 0.0 and " <>
+      "less than or equal to 1.0. Found #{inspect(size)}"
+  end
+
+  defp location_error(dim, size) do
+    "#{dim} must be a percentage expressed as a float beteen -1.0 and 1.0. " <>
+      "Found #{inspect(size)}"
   end
 
   defp invalid_crop(crop) do