+++ test/crop_test.exs
@@ -40,4 +40,30 @@ defmodule Image.Crop.Test do
 
     assert_images_equal(cropped, validate_path)
   end
+
+  test "Crop using percentages", %{image: image} do
+    assert Image.shape(Image.crop!(image, 0.0, 0.0, 1.0, 1.0)) == Image.shape(image)
+    assert Image.shape(Image.crop!(image, 0.0, 0.0, 0.5, 1.0)) == {500, 542, 3}
+    assert Image.shape(Image.crop!(image, 0.0, 0.0, 1.0, 0.5)) == {1000, 271, 3}
+
+    assert Image.shape(Image.crop!(image, 0.5, 0.5, 0.5, 0.5)) == {500, 271, 3}
+    assert Image.shape(Image.crop!(image, -0.5, 0.5, 0.5, 0.5)) == {500, 271, 3}
+    assert Image.shape(Image.crop!(image, 0.5, -0.5, 0.5, 0.5)) == {500, 271, 3}
+  end
+
+  test "Crop using invalid percentages", %{image: image} do
+    assert Image.crop(image, 0.5, 0.5, 0.6, 0.5) ==
+             {:error, "operation build: extract_area: bad extract area"}
+
+    assert Image.crop(image, 0.5, 0.5, 0.5, 0.6) ==
+             {:error, "operation build: extract_area: bad extract area"}
+
+    assert Image.crop(image, 0.5, 0.5, -0.5, 0.6) ==
+             {:error,
+              "width must be a percentage expressed as a float greater than 0.0 and less than or equal to 1.0. Found -0.5"}
+
+    assert Image.crop(image, 0.5, 0.5, 0.5, -0.6) ==
+             {:error,
+              "height must be a percentage expressed as a float greater than 0.0 and less than or equal to 1.0. Found -0.6"}
+  end
 end